requirements.txt
streamlit
plotly
scikit-learn
numpy
pandas
pillow
